Create README.md
This is a project I did for the artistic portion of an English assignment :) 

More in depth explanation:

I made this website as a part of the creative segment of an English asignment I had been assigned. It has a lot of references in it to the Crucible and the show Gossip Girls and I made it in a way to address the prompt of the project so let me give you some background information so you can further understand it! Basically we were trying to show how the Crucible (a movie and play about the Salem Witch Trials) compares to events that go on in the present. I though since there was lots of bullying and gossip that took place in the Crucible it reminded me of Gossip Girl which is a show that revolves around this website called Gossip Girl which people put gossip on. So I made a version of the Gossip Girl website, using references from the show and the crucibe to show how the Crucible was similar to Cyber Bullying.